Prevent home-redirect on refresh.
FlowRouter triggersEnter was relying on a buggy helper (also deleted) that was querying subscription data with no mechanism for requesting it.

This means that a user that refreshes would temporarily not be subscribed to a game room, and this logic would incorrectly think that meant the player is not in the room. The logic is now handled inside of the template, which correctly waits for the subscription (by existing implementation).