Fix proposer pop-up UI bugs
Moving between phases of the game, sometimes the previous proposer would see a gray screen overlay as if the pop-up was re-rendering but got interrupted. This was probably due to a delay between game phase change (which would newly render the pop-up) and the proposer change -- since these are two separate db updates.